generate signing keys on boot
a step toward using the tor key files
which I still can't figure out how to read and pass to RbNaCl

This struck me as a great step though,
in that we no longer need a two request send <-> fetch webhook cycle.

We can just start signing requests on the fly as hoped.